requirements.txt
asgiref==3.7.2
beautifulsoup4==4.12.3
boto3==1.34.44
botocore==1.34.44
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.3
defusedxml==0.7.1
dj-database-url==2.1.0
Django==4.2.7
django-allauth==0.61.1
django-bootstrap-v5==1.0.11
django-storages==1.14.2
gunicorn==21.2.0
idna==3.6
Jinja2==3.1.3
jmespath==1.0.1
MarkupSafe==2.1.5
oauthlib==3.2.2
packaging==23.2
psycopg2-binary==2.9.9
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python3-openid==3.2.0
PyYAML==6.0.1
render==1.0.0
requests==2.31.0
requests-oauthlib==1.3.1
s3transfer==0.10.0
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.0.7
whitenoise==6.6.0
django-crispy-forms
crispy-bootstrap5
